rename TransactionResult to FinalTransactionResult
