+++ src/XRP/index.ts
@@ -1,5 +1,5 @@
 export { default as PaymentFlags } from './model/payment-flags'
-export { default as TransactionStatus } from './transaction-status'
+export { default as TransactionStatus } from './final-transaction-status'
 export { default as XrpClient } from './xrp-client'
 export { default as XrpUtils } from './xrp-utils'
 export { default as XrpError, XrpErrorType } from './xrp-error'
