+++ src/XRP/xrp-client.ts
@@ -1,13 +1,13 @@
 import { Wallet, XrplNetwork } from 'xpring-common-js'
 import { BigInteger } from 'big-integer'
 import XrpClientDecorator from './xrp-client-decorator'
-import TransactionStatus from './transaction-status'
+import TransactionStatus from './final-transaction-status'
 import ReliableSubmissionXrpClient from './reliable-submission-xrp-client'
 import XrpClientInterface from './xrp-client-interface'
 import XrpTransaction from './model/xrp-transaction'
 
 import SendXrpDetails from './model/send-xrp-details'
-import TransactionResult from './model/transaction-result'
+import FinalTransactionResult from './model/transaction-result'
 
 /**
  * XrpClient is a client which interacts with the XRP Ledger.
@@ -144,7 +144,7 @@ export default class XrpClient implements XrpClientInterface {
    * @returns A promise which resolves to a TransactionResult object that contains the hash of the submitted AccountSet transaction,
    *          the final status of the transaction, and whether the transaction was included in a validated ledger.
    */
-  enableDepositAuth(wallet: Wallet): Promise<TransactionResult> {
+  enableDepositAuth(wallet: Wallet): Promise<FinalTransactionResult> {
     return this.decoratedClient.enableDepositAuth(wallet)
   }
 }
